+++ packages/expo-sqlite/CHANGELOG.md
@@ -23,6 +23,7 @@ _This version does not introduce any user-facing changes._
 ### 💡 Others
 
 - Updated API doc for `loadExtensionAsync()` and `loadExtensionSync()`. ([#38905](https://github.com/expo/expo/pull/38905) by [@kudo](https://github.com/kudo))
+- Change Constants to Constant/Property.
 
 ## 16.0.2 — 2025-08-16